Chỉnh lại bbox của mario lúc cầm mai rùa và lúc thả mai. Sửa lỗi mai rùa đi xuyên vật thể lúc thả. Mario sẽ bất tử một lúc sau khi thả mai

+++ 05-SceneManager/ShootingPlantHead.h
@@ -15,7 +15,7 @@ private:
 	// 2‑TopLeft
 	// 3‑BottomRight
 public:
-	CShootingPlantHead(int x, int y) : CGameObject(x, y) 
+	CShootingPlantHead(float x, float y) : CGameObject(x, y) 
 	{
 		z = 2;
 		this->orient = 0;
